Move EnumEntryJsonFormat to common package for reusability in V1 flow
